Properly structure pages and components
* Objective

This commit objective is about doing an initial folder structure for
pages and contents, as we are going to start growing the number of pages
for Encinitas.

* Why

We were about to create the second page (Transactions), and we had all
the components in a single file, this commit creates an initial (not
ideal) folder structure for Encinitas.

* How

This commit moves all the existing components and pages into sub-folders
with components and pages
