+++ src/App.tsx
@@ -1,84 +1,23 @@
 import * as React from 'react';
 import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
+
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
 import CssBaseline from '@mui/material/CssBaseline';
-import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import CardHeader from '@mui/material/CardHeader';
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import InfoIcon from '@mui/icons-material/Info';
-import Typography from '@mui/material/Typography';
 import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
 import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import Link from '@mui/material/Link';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-
-import { mainListItems, secondaryListItems } from './listItems';
-import ChartPerformance from './ChartPerformance';
-import ChartApdex from './ChartApdex';
-import ChartThroughput from './ChartThroughput';
-import ChartErrorRate from './ChartErrorRate';
 
 import theme from './theme';
+import AppBar from './components/AppBar';
+import Drawer from './components/Drawer';
 
-const drawerWidth: number = 240;
-
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean;
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})<AppBarProps>(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    '& .MuiDrawer-paper': {
-      position: 'relative',
-      whiteSpace: 'nowrap',
-      width: drawerWidth,
-      transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      boxSizing: 'border-box',
-      ...(!open && {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-          width: theme.spacing(9),
-        },
-      }),
-    },
-  }),
-);
+import Dashboard from './pages/Dashboard/Dashboard';
+import Transactions from './pages/Transactions/Transactions';
 
 export default function App() {
   const [open, setOpen] = React.useState(true);
@@ -88,155 +27,30 @@ export default function App() {
 
   return (
     <ThemeProvider theme={theme}>
-      <Box sx={{ display: 'flex' }}>
-        <CssBaseline />
-        <AppBar position="absolute" open={open}>
-          <Toolbar
+      <Router>
+        <Box sx={{ display: 'flex' }}>
+          <CssBaseline />
+          <AppBar open={open} toggleDrawer={toggleDrawer} />
+          <Drawer open={open} toggleDrawer={toggleDrawer} />
+          <Box
+            component="main"
             sx={{
-              pr: '24px', // keep right padding when drawer closed
+              marginTop: '40px',
+              backgroundColor: (theme) =>
+                theme.palette.mode === 'light'
+                  ? theme.palette.grey[100]
+                  : theme.palette.grey[900],
+              flexGrow: 1,
+              overflow: 'auto',
             }}
           >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: '36px',
-                ...(open && { display: 'none' }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Dashboard
-            </Typography>
-            <IconButton color="inherit">
-              <Badge badgeContent={4} color="secondary">
-                <NotificationsIcon />
-              </Badge>
-            </IconButton>
-          </Toolbar>
-        </AppBar>
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              px: [1],
-            }}
-          >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
-            </IconButton>
-            <Typography
-              variant="h4"
-              noWrap
-              component="div"
-              sx={{
-                flexGrow: 1,
-                color: '#979797',
-                fontFamily: 'Oswald, sans-serif', fontWeight: 'bold'
-              }}
-            >
-              ENCINITAS
-            </Typography>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            {mainListItems}
-            <Divider sx={{ my: 1 }} />
-            {secondaryListItems}
-          </List>
-        </Drawer>
-        <Box
-          component="main"
-          sx={{
-            backgroundColor: (theme) =>
-              theme.palette.mode === 'light'
-                ? theme.palette.grey[100]
-                : theme.palette.grey[900],
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
-          }}
-        >
-          <Toolbar />
-          <Container maxWidth={false} sx={{ mt: 4, mb: 4 }}>
-            <Grid container spacing={3}>
-              <Grid item xs={12} md={6} lg={6}>
-                <Card sx={{ borderRadius: '16px', boxShadow: 3, maxWidth: '100%', mb: 3 }}>
-                  <CardHeader
-                    title={
-                      <Typography variant="h6" component="div" sx={{ display: 'flex', alignItems: 'center' }}>
-                        dApp Application Performance
-                        <InfoIcon sx={{ fontSize: '1.0em', ml: 1 }} /> {}
-                      </Typography>
-                    }
-                  />
-                  <CardContent>
-                    <ChartPerformance />
-                  </CardContent>
-                </Card>
-              </Grid>
-              <Grid item xs={12} md={6} lg={6}>
-                <Card sx={{ borderRadius: '16px', boxShadow: 3, maxWidth: '100%'}}>
-                  <CardHeader
-                    title={
-                      <Typography variant="h6" component="div" sx={{ display: 'flex', alignItems: 'center' }}>
-                        User Satisfaction - Apdex Score
-                        <InfoIcon sx={{ fontSize: '1.0em', ml: 1 }} /> {}
-                      </Typography>
-                    }
-                  />
-                  <CardContent>
-                    <ChartApdex />
-                  </CardContent>
-                </Card>
-              </Grid>
-            </Grid>
-            <Grid container spacing={3}>
-              <Grid item xs={12} md={6} lg={6}>
-                <Card sx={{ borderRadius: '16px', boxShadow: 3, maxWidth: '100%', mb: 3}}>
-                <CardHeader
-                  title={
-                    <Typography variant="h6" component="div" sx={{ display: 'flex', alignItems: 'center' }}>
-                      Throughput
-                      <InfoIcon sx={{ fontSize: '1.0em', ml: 1 }} /> {}
-                    </Typography>
-                  }
-                />
-                  <CardContent>
-                    <ChartThroughput />
-                  </CardContent>
-                </Card>
-              </Grid>
-              <Grid item xs={12} md={6} lg={6}>
-                <Card sx={{ borderRadius: '16px', boxShadow: 3, maxWidth: '100%'}}>
-                  <CardHeader
-                    title={
-                      <Typography variant="h6" component="div" sx={{ display: 'flex', alignItems: 'center' }}>
-                        Error Rate
-                        <InfoIcon sx={{ fontSize: '1.0em', ml: 1 }} /> {}
-                      </Typography>
-                    }
-                  />
-                  <CardContent>
-                    <ChartErrorRate />
-                  </CardContent>
-                </Card>
-              </Grid>
-            </Grid>
-          </Container>
+            <Routes>
+              <Route path="/" element={<Dashboard />} />
+              <Route path="/transactions" element={<Transactions />} />
+            </Routes>
+          </Box>
         </Box>
-      </Box>
+      </Router>
     </ThemeProvider>
   );
 }
