+++ src/components/ChartPerformance.tsx
