+++ src/components/ChartThroughput.tsx
