+++ src/components/Drawer.tsx
@@ -1,0 +1,88 @@
+import * as React from 'react';
+
+import { styled } from '@mui/material/styles';
+
+import Divider from '@mui/material/Divider';
+import IconButton from '@mui/material/IconButton';
+import MuiDrawer from '@mui/material/Drawer';
+import List from '@mui/material/List';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
+
+import { mainListItems, secondaryListItems } from './drawerItems';
+
+// TODO: Move to theme
+const drawerWidth: number = 240;
+
+interface DrawerProps {
+  open?: boolean;
+  toggleDrawer?: () => void;
+}
+
+const DrawerComponent = styled(MuiDrawer, {
+  shouldForwardProp: (prop) => prop !== 'open' })(
+  ({ theme, open }) => ({
+    '& .MuiDrawer-paper': {
+      position: 'relative',
+      whiteSpace: 'nowrap',
+      width: drawerWidth,
+      transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+      }),
+      boxSizing: 'border-box',
+      ...(!open && {
+        overflowX: 'hidden',
+        transition: theme.transitions.create('width', {
+          easing: theme.transitions.easing.sharp,
+          duration: theme.transitions.duration.leavingScreen,
+        }),
+        width: theme.spacing(7),
+        [theme.breakpoints.up('sm')]: {
+          width: theme.spacing(9),
+        },
+      }),
+    },
+  }),
+);
+
+export default function Drawer(
+  { open, toggleDrawer }: DrawerProps
+) {
+return (
+  <DrawerComponent variant="permanent" open={open}>
+    <Toolbar
+      sx={{
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'flex-end',
+        px: [1],
+      }}
+    >
+      <IconButton onClick={toggleDrawer}>
+        <ChevronLeftIcon />
+      </IconButton>
+      <Typography
+        variant="h4"
+        noWrap
+        component="div"
+        sx={{
+          flexGrow: 1,
+          color: '#979797',
+          fontFamily: 'Oswald, sans-serif', fontWeight: 'bold'
+        }}
+      >
+        ENCINITAS
+      </Typography>
+    </Toolbar>
+    <Divider />
+    <List component="nav">
+      {mainListItems}
+      <Divider sx={{ my: 1 }} />
+      {secondaryListItems}
+    </List>
+  </DrawerComponent>
+);
+}
