+++ src/pages/Transactions/Transactions.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export default function Transactions() {
+return <div>Transactions</div>;
+}
